Moved function 'get_module_list()' to utils
* Function is useful not just for building the documentation
