+++ test/utils/moduleinterfaces/test_moduleinterfaces.py
@@ -9,17 +9,17 @@ def test_basic():
     cpacs_inout = CPACSInOut()
 
     cpacs_inout.add_input(
-            descr='Test description',
-            cpacs_path='/cpacs/testpath',
-            default_value=5,
-            unit='m/s',
-            var_name=None
-        )
+        descr='Test description',
+        cpacs_path='/cpacs/testpath',
+        default_value=5,
+        unit='m/s',
+        var_name=None
+    )
 
     cpacs_inout.add_output(
-            descr='Test description',
-            cpacs_path='/cpacs/testpath',
-            default_value=None,
-            unit='m/s',
-            var_name=None
-        )
+        descr='Test description',
+        cpacs_path='/cpacs/testpath',
+        default_value=None,
+        unit='m/s',
+        var_name=None
+    )
